+++ packages/fireworks/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ai-sdk/fireworks
 
+## 3.0.0-canary.44
+
+### Patch Changes
+
+- Updated dependencies [d848405]
+  - @ai-sdk/provider-utils@5.0.0-canary.37
+  - @ai-sdk/openai-compatible@3.0.0-canary.44
+
+## 3.0.0-canary.43
+
+### Patch Changes
+
+- Updated dependencies [ca39020]
+  - @ai-sdk/provider-utils@5.0.0-canary.36
+  - @ai-sdk/openai-compatible@3.0.0-canary.43
+
 ## 3.0.0-canary.42
 
 ### Patch Changes
@@ -719,13 +735,13 @@
   Before
 
   ```ts
-  model.textEmbeddingModel("my-model-id");
+  model.textEmbeddingModel('my-model-id');
   ```
 
   After
 
   ```ts
-  model.embeddingModel("my-model-id");
+  model.embeddingModel('my-model-id');
   ```
 
 - 95f65c2: chore: use import \* from zod/v4
@@ -953,13 +969,13 @@
   Before
 
   ```ts
-  model.textEmbeddingModel("my-model-id");
+  model.textEmbeddingModel('my-model-id');
   ```
 
   After
 
   ```ts
-  model.embeddingModel("my-model-id");
+  model.embeddingModel('my-model-id');
   ```
 
 - Updated dependencies [8d9e8ad]
@@ -1447,7 +1463,7 @@
 
   ```js
   await generateImage({
-    model: luma.image("photon-flash-1", {
+    model: luma.image('photon-flash-1', {
       maxImagesPerCall: 5,
       pollIntervalMillis: 500,
     }),
@@ -1460,7 +1476,7 @@
 
   ```js
   await generateImage({
-    model: luma.image("photon-flash-1"),
+    model: luma.image('photon-flash-1'),
     prompt,
     n: 10,
     maxImagesPerCall: 5,
@@ -1748,7 +1764,7 @@
 
   ```js
   await generateImage({
-    model: luma.image("photon-flash-1", {
+    model: luma.image('photon-flash-1', {
       maxImagesPerCall: 5,
       pollIntervalMillis: 500,
     }),
@@ -1761,7 +1777,7 @@
 
   ```js
   await generateImage({
-    model: luma.image("photon-flash-1"),
+    model: luma.image('photon-flash-1'),
     prompt,
     n: 10,
     maxImagesPerCall: 5,