+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,36 @@
 
 ### Patch Changes
 
+- ai@7.0.0-canary.130
+
+## 0.0.1-canary.0
+
+### Patch Changes
+
+- Updated dependencies [d1b3786]
+  - ai@7.0.0-canary.129
+
+## 0.0.1-canary.0
+
+### Patch Changes
+
+- ai@7.0.0-canary.128
+
+## 0.0.1-canary.0
+
+### Patch Changes
+
+- Updated dependencies [e95e38d]
+- Updated dependencies [016e877]
+- Updated dependencies [ca99fea]
+- Updated dependencies [d775a57]
+- Updated dependencies [538c12b]
+  - ai@7.0.0-canary.127
+
+## 0.0.1-canary.0
+
+### Patch Changes
+
 - ai@7.0.0-canary.126
 
 ## 0.0.1-canary.0